Add files via upload
Recent saliency detection methods have made great progress with the fully convolutional network. However, we find that
the saliency maps are usually coarse and fuzzy, especially near the boundary of the salient object. To deal with this problem,
in this paper, we exploit a multi-path feature fusion model for saliency detection. The proposed model is a fully convolutional
network with raw images as input and saliency maps as output. In particular, we propose a multi-path fusion strategy
for deriving the intrinsic features of salient objects. The structure has the ability to capture the low-level visual features
and generating the boundary-preserving saliency maps. Moreover, a coupled structure module is proposed in our
model, which helps to explore the high-level semantic properties of salient objects. Extensive experiments on four public
benchmarks indicate that our saliency model is effective and outperforms state-of-the-art methods.

Our results
You can also download from the link：https://pan.baidu.com/s/1eWRB08QPx9FbYRwLaYl0aQ